- Terminada la parte de actualizar los menus cuando se hace login - Mejorada la perspectiva para no mostrar el editor - Las acciones estan en una interfaz y coinciden con los nombres del XML - Extendida la calse GraphViewer para poder acceder al grafo y añadir listeners. - Las rutas de los archivos de configuracion de la base de datos y los XML se han puesto en una interfaz

+++ plugin/gcad/src/model/business/knowledge/Session.java
@@ -14,7 +14,7 @@ public class Session implements ISession {
 		return idSesion;
 	}
 
-	public long getRol() {
+	public int getRol() {
 		return user.getRol().ordinal();
 	}
